Fix SonarLint reported vulnerabilities.
Get the temp directory from system properties. This ensures correct behavior on systems where /tmp is not the correct directory.
Correctly log an exception in CsvImportLayout.java.
Made an array of strings in FiltersLayout.java package private and final. These were never set and were used as constants within the package.
In CsvImportLayout.isNumberString catch only the exception that can be thrown.