#702: Fix presence of NUL characters/empty bytes in various export types (SKOS, PDF, HTML)
AccessibleByteArrayOutputStream used to return the entire backing buffer, but because the buffer is often larger than the data stored within it results in the end of the buffer being empty (or a sequence of NUL characters in ASCII). This is harmless for most export types as the empty bytes won't be referenced, but it does make downloads bigger than they should be.