#346: Fix various Java compiler warnings
Most of the warnings were caused by Serializable classes missing serialVersionUID fields. Others were caused by missing Mapstruct mappings. Some Mapstruct warnings are still emitted as there is a case to be made that the mappings are incomplete.

CodeAlreadyExistException and VocabularyAlreadyExistException have been modified so that the code or the vocabulary that exists is now reported.