Fix agency and licence image uploads not working
The CVS server returns the location of the uploaded images in the HTTP location header. The CVS client attempted to check whether a body was present, as per previous versions of the CVS server.

The code for handling uploads moved to FileUploadService to reduce code duplication.

The proxy configuration was also updated so that dev server uploads were also fixed.