Fix rendering of close buttons in modal dialogs
Due to Bootstrap upgrades, the previous method of rendering buttons doesn't look correct. This commit replaces all modal close buttons with Bootstrap native buttons.

In many dialogs there is two close buttons, one in the header and another in the footer. We should consider which ones to keep in a separate issue.