#989: Remove unused edit interface
The edit interface removed could allow changes to be made to the database outside of the enforcement mechanisms. This caused issues when trying to resolve #1070 when these interfaces were used.

This commit also refactors the roles system to use the enum, increasing type safety.

This commit fixes access denied changing the page URL.

This closes #989