#1095: Fix `/v2/vocabularies/{vocabulary}/{versionNumberSl}` redirecting to the web application when `text/html` is requested, remove `application/xhtml+xml` support from the `/v2/vocabularies/{vocabulary}/{versionNumberSl}` endpoint
The original `/v2/vocabularies/{vocabulary}/{versionNumberSl}` endpoint for `text/html` was a simple redirect controller. The frontend used to use specific `/v2/vocabularies/{format}/{vocabulary}/{versionNumberSl}` endpoints for vocabulary export, but this was changed as of #958. This behavior was not explained in the API documentation. This commit removes the redirect controller and returns the export.

This commit also removes `application/xhtml+xml` from the `/v2/vocabularies/{vocabulary}/{versionNumberSl}` as it doesn't return valid XML. It also removes the format specific endpoints as these are unused and are not part of the public API.

This closes #1095